working building system

+++ include/Object/Bush.hpp
@@ -9,7 +9,7 @@ public:
     Bush(sf::Vector2f position) : WorldObject(position) {}
 
     inline void update(float dt) override;
-    void draw(sf::RenderWindow& window, float dt, float alpha) override;
+    void draw(sf::RenderWindow& window, float dt, int alpha) override;
 
     inline void interact() override;
